di set

+++ app/src/main/java/com/bmprj/cointicker/ui/detail/CoinDetailViewModel.kt
@@ -1,32 +1,30 @@
 package com.bmprj.cointicker.ui.detail
 
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.bmprj.cointicker.data.remote.coin.CoinRepositoryImpl
-import com.bmprj.cointicker.data.remote.firebase.auth.AuthRepository
 import com.bmprj.cointicker.data.remote.firebase.cloud.CloudRepository
 import com.bmprj.cointicker.domain.coins.CoinDetailEntity
 import com.bmprj.cointicker.domain.coins.GetCoinUseCase
-import com.bmprj.cointicker.utils.Resource
 import com.bmprj.cointicker.model.CoinDetail
 import com.bmprj.cointicker.model.FavouriteCoin
 import com.bmprj.cointicker.utils.UiState
+import com.google.firebase.auth.FirebaseUser
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.launch
+import javax.annotation.Nullable
 import javax.inject.Inject
 
 @HiltViewModel
 class CoinDetailViewModel @Inject constructor(
-    private val authRepository: AuthRepository,
     private val coinUseCase: GetCoinUseCase,
-    private val cloudRepository: CloudRepository
+    private val cloudRepository: CloudRepository,
+    @Nullable private val firebaseUser:FirebaseUser?
 ) :ViewModel(){
+
 
     private val _coinDetail = MutableStateFlow<UiState<CoinDetailEntity>>(UiState.Loading)
     val coinDetail = _coinDetail.asStateFlow()
@@ -37,14 +35,19 @@ class CoinDetailViewModel @Inject constructor(
     private val _favouriteDelete = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
     val favouriteDelete = _favouriteDelete
 
-    val _isFavourite = MutableLiveData<UiState<Boolean>>()
+    private val _isFavourite = MutableStateFlow<UiState<Boolean>>(UiState.Loading)
+    val isFavourite = _isFavourite.asStateFlow()
+
+    val currentUserId = firebaseUser?.uid
 
 
-    val currentUserId = authRepository.currentUser?.uid!!
+    fun setFavouriteState(is_favourite:Boolean) = viewModelScope.launch {
+        _isFavourite.emit(UiState.Success(is_favourite))
+    }
 
 
     fun addFavourite(coinDetaill: CoinDetail) = viewModelScope.launch {
-
+        if (currentUserId == null) return@launch
         val favList = FavouriteCoin(coinDetaill.id, coinDetaill.name, coinDetaill.image.small, coinDetaill.symbol)
 
         cloudRepository.addFavourite(currentUserId,favList)
@@ -61,7 +64,7 @@ class CoinDetailViewModel @Inject constructor(
     }
 
     fun getFavourite(coinId: String) = viewModelScope.launch {
-
+        if (currentUserId == null) return@launch
         cloudRepository.getFavourite(currentUserId,coinId)
             .onStart {
                 _isFavourite.value = UiState.Loading
@@ -75,7 +78,7 @@ class CoinDetailViewModel @Inject constructor(
     }
 
     fun delete(coinId:String) = viewModelScope.launch {
-
+        if (currentUserId == null) return@launch
         cloudRepository.delete(currentUserId,coinId)
             .onStart {
                 _favouriteDelete.emit(UiState.Loading)
@@ -89,7 +92,6 @@ class CoinDetailViewModel @Inject constructor(
     }
 
     fun getCoin(id:String) = viewModelScope.launch {
-
         coinUseCase.getCoin(id)
             .onStart {
                 _coinDetail.emit(UiState.Loading)
